refactor(wip): move native wrapper credential preparation function to lib/native-wrapper
