+++ src/lib/native-wrapper/index.ts
@@ -1,0 +1,1 @@
+export { prepareCredentialsForNativeWrapper } from './credentials';
